# task/com.achelos.task.rest-impl/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: "TaSK Framework - Remote API"
  description: "This document defines a RESTful API for executing runs of the TaSK testsuites and accessing their result."
  version: "1.0.0"
  license:
    name: "European Union Public Licence (EUPL) 1.2"
    url: "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
servers:
  - url: '{protocol}://{task_ip}:{task_port}/'
    description: The TaSK REST API.
    variables:
      protocol:
        default: https
        enum:
          - http
          - https
      task_ip:
        default: 127.0.0.1
        description: IP or HostName of TaSK Server, which is served at the location the OpenAPI YAML is served.
      task_port:
        default: '8088'
        description: Port of TaSK Server.
...
paths:
  /:
    get:
      tags:
        - "Base Path"
      summary: "Return the API definition of the REST Server."
      description: "Return the API definition of the REST Server."
      operationId: getAPIDefinition
      responses:
        200:
          $ref: '#/components/responses/ApiDefinition'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /execute/mics:
    post:
      tags:
        - "Start Execution"
      summary: "Execute a Testrun of the TaSK Framework using a MICS file."
      description: "Execute the TaSK Framework with the provided MICS file."
      operationId: executeFromMICS
      requestBody:
        $ref: '#/components/requestBodies/MICS'
      responses:
        202:
          $ref: '#/components/responses/RunId'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
        503:
          $ref: '#/components/responses/ErrorResponse'
  /execute/trp:
    post:
      tags:
        - "Start Execution"
      summary: "Execute a Testrun of the TaSK Framework from a given Test Runplan."
      description: "Execute a Testrun of the TaSK Framework with the provided Test Runplan file."
      operationId: executeFromRunplan
      requestBody:
        $ref: '#/components/requestBodies/TestRunplan'
      responses:
        202:
          $ref: '#/components/responses/RunId'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
        503:
          $ref: '#/components/responses/ErrorResponse'
...
components:
...
  requestBodies:
    TestRunplan:
      description: "Information regarding the TestRunplan of the Device under Test and related information."
      required: true
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/schemas/TestRunPlan'
    MICS:
      description: "Information regarding the MICS of the Device under Test and related information."
      required: true
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/schemas/MICS'
  responses:
    ApiDefinition:
      description: "OK"
      content:
        text/yaml:
          schema:
            $ref: '#/components/schemas/APIDefinition'
    RunId:
      description: "Accepted"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RunId'
...
    ErrorResponse:
      description: "ErrorMessage"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    APIDefinition:
      type: string
    TestRunPlan:
      type: object
      properties:
        testRunPlanFile:
          type: string
          format: binary
          description: "The TestRunplan file itself."
        clientAuthCertificateChain:
          description: "The certificate chain a test client should use to verify client authentication."
          type: array
          items:
            type: string
            format: binary
        clientAuthKey:
          description: "The private key a test client should use to verify client authentication."
          type: string
          format: binary
    MICS:
      type: object
      properties:
        micsFile:
          type: string
          format: binary
          description: "The MICS file itself."
        serverCertificateChain:
          type: array
          description: "The certificate chain of the server under test, from which the fingerprints are specified in the MICS file."
          items:
            type: string
            format: binary
        ignoreMicsVerificationResult:
          type: boolean
          description: "Ignore the result of the MICS verification when running the resulting test cases."
        clientAuthCertificateChain:
          description: "The certificate chain a test client should use to verify client authentication."
          type: array
          items:
            type: string
            format: binary
        clientAuthKey:
          description: "The private key a test client should use to verify client authentication."
          type: string
          format: binary
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        statuscode:
          type: string
        errormessage:
          type: string
        path:
          type: string
...
    RunId:
      type: object
      properties:
        runId:
          type: string
          example: "a4b4d23234b23ef23a"
...